✨ feat(user): user service/handler updated to format the API response with the expected format

+++ internal/user/repository/user_repository.go
@@ -8,6 +8,6 @@ import (
 
 type UserRepository interface {
 	GetAll(ctx context.Context) ([]domain.User, error)
-	GetByID(ctx context.Context, id int) (*domain.User, error)
+	GetByID(ctx context.Context, id int64) (*domain.User, error)
 	Create(ctx context.Context, user *domain.User) (*domain.User, error)
 }
